requirements.txt
# ─── Aura — Python Dependencies ───────────────────────────────────────────────
# Install: pip install -r requirements.txt

# ── Web Framework & Server ────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── Streamlit UI ──────────────────────────────────────────────────────────────
streamlit==1.35.0

# ── LangGraph Agent ───────────────────────────────────────────────────────────
langgraph==0.1.14
langchain-core==0.2.5

# ── Auth (JWT) ────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── Data & Forecasting ────────────────────────────────────────────────────────
pydantic==2.7.3
pandas==2.2.2
python-dateutil==2.9.0

# ── Networking ────────────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.27.0

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
